requirements.txt
# Core Flask
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.50

# Authentication & Security
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
email-validator==2.1.0.post1

# Real-time Communication
Flask-SocketIO==5.3.5
python-socketio==5.7.2
eventlet==0.33.3
gevent==23.9.1
gevent-websocket==0.10.1

# Internationalization
Flask-Babel==4.0.0
Babel==2.13.1

# Forms & Validation
Flask-WTF==1.2.2
WTForms==3.0.1
email-validator==2.1.0.post1

# Caching
Flask-Caching==2.1.0
redis==5.0.1  # For production caching

# Payments
razorpay==1.4.2
stripe==7.6.0  # Alternative payment processor

# File Processing
Pillow==10.1.0  # For image processing
python-magic==0.4.27  # For file type validation

# Email
Flask-Mail==0.9.1

# Monitoring & Logging
sentry-sdk[flask]==1.34.0
gunicorn==21.2.0  # Production WSGI server

# Development Tools
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Production Dependencies
python-dotenv==1.0.0  # For loading .env files
python-dateutil==2.8.2
pytz==2023.3.post1
